+++ adapters/mqtt-vertx/src/main/java/org/eclipse/hono/adapter/mqtt/impl/Application.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2017 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2020 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -24,6 +24,7 @@ import org.springframework.context.annotation.Configuration;
  */
 @ComponentScan("org.eclipse.hono.adapter.mqtt")
 @ComponentScan("org.eclipse.hono.service.metric")
+@ComponentScan("org.eclipse.hono.deviceconnection.infinispan.client")
 @Configuration
 @EnableAutoConfiguration
 public class Application extends AbstractApplication {